Add a test to cover edge cases
