+++ src/test/packages/31-component-actions-edgecases/zarf.yaml
@@ -1,0 +1,17 @@
+kind: ZarfPackageConfig
+metadata:
+  name: component-actions-edgecases
+  description: Component actions edge cases
+
+components:
+  - name: on-deploy
+    required: true
+    actions:
+      # runs during "zarf package remove"
+      onDeploy:
+        before:
+          # Wait for something that doesn't have a namespace (no -n)
+          - wait:
+              cluster:
+                kind: ns
+                name: zarf
